# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Newsletters API",
        "description": "Newsletters API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "description": "Read all the articles in the database",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/articles/store": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "description": "Create a new article",
                "operationId": "store",
                "requestBody": {
                    "description": "Pass article data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image_placement"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Article Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Article Description"
                                    },
                                    "image_placement": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "left"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/articles/show/{ArticleID}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by id",
                "description": "Read article by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "ArticleID",
                        "in": "path",
                        "description": "ID of article to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/articles/edit/{ArticleID}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by id",
                "description": "Read article by id",
                "operationId": "edit",
                "parameters": [
                    {
                        "name": "ArticleID",
                        "in": "path",
                        "description": "ID of article to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/articles/update/{ArticleID}": {
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update article by id",
                "description": "Update article by id",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "ArticleID",
                        "in": "path",
                        "description": "ID of article to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass article data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image_placement"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Article Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Article Description"
                                    },
                                    "image_placement": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "left"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/articles/destroy/{ArticleID}": {
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete article by id",
                "description": "Delete article by id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "ArticleID",
                        "in": "path",
                        "description": "ID of article to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        }
    }
}